Update: comment spruce ((param))s in skeleton files
The current setup of the pipeline heavily relies on the ((param)) feature of spruce in conjunction with --skip-eval. Spruce does not yet support the two features in combination: https://github.com/geofffranks/spruce/issues/211
The workaround chosen is to just comment the ((param)) usages, which leaves functionality intact, but hinders spruce from automatically detecting semantical mistakes in the config files.